+++ CPR/CprServiceInterface.php
@@ -10,7 +10,8 @@
 
 namespace Kontrolgruppen\CoreBundle\CPR;
 
-use Kontrolgruppen\CoreBundle\Entity\Client;
+use Kontrolgruppen\CoreBundle\Entity\AbstractProcessClient;
+use Kontrolgruppen\CoreBundle\Entity\ProcessClientPerson;
 
 /**
  * Interface CprServiceInterface.
@@ -30,22 +31,22 @@ interface CprServiceInterface
      * Populates client with information from the CPR service. If no data is found via the service the client
      * object is returned without being changed.
      *
-     * @param Cpr    $cpr
-     * @param Client $client
+     * @param Cpr                 $cpr
+     * @param ProcessClientPerson $client
      *
-     * @return Client
+     * @return ProcessClientPerson
      *
      * @throws CprException
      */
-    public function populateClient(Cpr $cpr, Client $client): Client;
+    public function populateClient(Cpr $cpr, ProcessClientPerson $client): ProcessClientPerson;
 
     /**
-     * @param Cpr    $cpr
-     * @param Client $client
+     * @param Cpr                   $cpr
+     * @param AbstractProcessClient $client
      *
      * @return bool
      *
      * @throws CprException
      */
-    public function isNewClientInfoAvailable(Cpr $cpr, Client $client): bool;
+    public function isNewClientInfoAvailable(Cpr $cpr, ProcessClientPerson $client): bool;
 }