+++ Form/ChannelType.php
@@ -11,6 +11,7 @@
 namespace Kontrolgruppen\CoreBundle\Form;
 
 use Kontrolgruppen\CoreBundle\Entity\Channel;
+use Kontrolgruppen\CoreBundle\Form\Process\ClientTypesType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
@@ -26,6 +27,12 @@ class ChannelType extends AbstractType
      */
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
+        $builder
+            ->add('clientTypes', ClientTypesType::class, [
+                'label' => 'channel.form.client_types.label',
+                'help' => 'channel.form.client_types.help',
+            ]);
+
         $builder
             ->add('name', null, [
                 'label' => 'channel.form.name',