+++ Repository/AbstractProcessClientRepository.php
@@ -1,0 +1,35 @@
+<?php
+
+/*
+ * This file is part of aakb/kontrolgruppen-core-bundle.
+ *
+ * (c) 2019 ITK Development
+ *
+ * This source file is subject to the MIT license.
+ */
+
+namespace Kontrolgruppen\CoreBundle\Repository;
+
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\Common\Persistence\ManagerRegistry;
+
+/**
+ * Base class for process client repositories.
+ */
+class AbstractProcessClientRepository extends ServiceEntityRepository
+{
+    /**
+     * The process client class for this repository.
+     *
+     * @var string
+     */
+    protected static $class;
+
+    /**
+     * {@inheritdoc}
+     */
+    public function __construct(ManagerRegistry $registry)
+    {
+        parent::__construct($registry, static::$class);
+    }
+}