+++ Repository/ServiceRepository.php
@@ -10,11 +10,10 @@
 
 namespace Kontrolgruppen\CoreBundle\Repository;
 
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Persistence\ManagerRegistry;
 use Kontrolgruppen\CoreBundle\Entity\Process;
 use Kontrolgruppen\CoreBundle\Entity\ProcessType;
 use Kontrolgruppen\CoreBundle\Entity\Service;
+use Kontrolgruppen\CoreBundle\Repository\Traits\FindByProcessTypeTrait;
 
 /**
  * @method Service|null find($id, $lockMode = null, $lockVersion = null)
@@ -22,15 +21,14 @@ use Kontrolgruppen\CoreBundle\Entity\Service;
  * @method Service[]    findAll()
  * @method Service[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
  */
-class ServiceRepository extends ServiceEntityRepository
+class ServiceRepository extends AbstractTaxonomyRepository
 {
+    use FindByProcessTypeTrait;
+
     /**
      * {@inheritdoc}
      */
-    public function __construct(ManagerRegistry $registry)
-    {
-        parent::__construct($registry, Service::class);
-    }
+    protected static $taxonomyClass = Service::class;
 
     /**
      * Get services by process type.